# Backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Connect'in API",
        "description": "API documentation for Connect'in",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/createGroup": {
            "post": {
                "tags": [
                    "Groups"
                ],
                "summary": "Create a group",
                "operationId": "21cc46327f93ab78ee10be02345c7c39",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "My Group"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Group created"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/getGroups": {
            "get": {
                "tags": [
                    "Groups"
                ],
                "summary": "Get all groups of the authenticated user",
                "operationId": "dcaf9db67f9954086b2344e011344846",
                "responses": {
                    "200": {
                        "description": "List of groups"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/group/{group}": {
            "get": {
                "tags": [
                    "Groups"
                ],
                "summary": "Get a single group",
                "operationId": "7a45dab3abfe55ed20a3bbcf8177deab",
                "parameters": [
                    {
                        "name": "group",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Group details"
                    },
                    "404": {
                        "description": "Group not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/getAllUsers": {
            "get": {
                "tags": [
                    "Groups"
                ],
                "summary": "Get all users",
                "operationId": "6632d9aa8071cc56e96bd33d4b2182b4",
                "responses": {
                    "200": {
                        "description": "List of users"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}